requirements.txt
duckduckgo_search==7.2.1
openai==1.60.0
phidata==2.7.9